training/requirements.txt
torch>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
onnx>=1.16.0
onnxscript>=0.1.0
onnxruntime>=1.17.0
tqdm>=4.66.0
pyyaml>=6.0.1